+++ src/Services/WordCounter/WordCounter.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace PHPUnuhi\Services\WordCounter;
+
+class WordCounter
+{
+
+    /**
+     * @param string $text
+     * @return int
+     */
+    public function getWordCount(string $text): int
+    {
+        return str_word_count($text);
+    }
+
+}